Titles and Scoring
! Removed Selection and User from Title in Domain
+ Introduced Score, Award and Comment rights
+ UpdateScoreCategory
+ Updated Add operations with OperationResult to return 201 instead of 200
+ Fixed path parameters of Comments
+ Title can only be uniquely awarded to single Application
+ Comment now has User property in Domain
+ GetAllTitles
+ AddTitle
+ UpdateTitle
+ RemoveTitle
+ Updated Postman Collection

+++ src/Mvp.Selections.Api/Services/CommentService.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Net;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
@@ -35,11 +36,12 @@ namespace Mvp.Selections.Api.Services
                 ApplicationComment newComment = new (Guid.Empty)
                 {
                     Application = applicationResult.Result,
+                    User = user,
                     Value = comment.Value
                 };
                 result.Result = _commentRepository.Add(newComment) as ApplicationComment;
                 await _commentRepository.SaveChangesAsync();
-                result.StatusCode = HttpStatusCode.OK;
+                result.StatusCode = HttpStatusCode.Created;
             }
             else
             {
@@ -49,11 +51,11 @@ namespace Mvp.Selections.Api.Services
             return result;
         }
 
-        public async Task<OperationResult<Comment>> UpdateCommentAsync(Guid id, Comment comment)
+        public async Task<OperationResult<Comment>> UpdateCommentAsync(User user, Guid id, Comment comment)
         {
             OperationResult<Comment> result = new ();
             Comment existingComment = await _commentRepository.GetAsync(id);
-            if (existingComment != null)
+            if (existingComment != null && (existingComment.User.Id == user.Id || user.HasRight(Right.Admin)))
             {
                 existingComment.Value = comment.Value;
                 await _commentRepository.SaveChangesAsync();
@@ -61,9 +63,16 @@ namespace Mvp.Selections.Api.Services
                 result.Result = existingComment;
                 result.StatusCode = HttpStatusCode.OK;
             }
+            else if (existingComment != null)
+            {
+                string message = $"User '{user.Id}' tried to update Comment '{id}' but is not authorized to do so.";
+                _logger.LogWarning(message);
+                result.Messages.Add(message);
+                result.StatusCode = HttpStatusCode.Forbidden;
+            }
             else
             {
-                string message = $"Could not find Comment '{id}'.";
+                string message = $"Comment '{id}' not found.";
                 _logger.LogInformation(message);
                 result.Messages.Add(message);
             }
@@ -71,12 +80,47 @@ namespace Mvp.Selections.Api.Services
             return result;
         }
 
-        public async Task RemoveCommentAsync(Guid id)
+        public async Task<OperationResult<object>> RemoveCommentAsync(User user, Guid id)
         {
-            if (await _commentRepository.RemoveAsync(id))
+            OperationResult<object> result = new ();
+            Comment comment = await _commentRepository.GetAsync(id);
+            if (comment != null && (comment.User.Id == user.Id || user.HasRight(Right.Admin)))
             {
-                await _commentRepository.SaveChangesAsync();
+                if (await _commentRepository.RemoveAsync(id))
+                {
+                    await _commentRepository.SaveChangesAsync();
+                }
             }
+            else if (comment != null && comment.User.Id != user.Id)
+            {
+                string message = $"User '{user.Id}' tried to delete Comment '{id}' but is not authorized to do so.";
+                _logger.LogWarning(message);
+                result.Messages.Add(message);
+                result.StatusCode = HttpStatusCode.Forbidden;
+            }
+            else
+            {
+                result.StatusCode = HttpStatusCode.NoContent;
+            }
+
+            return result;
+        }
+
+        public async Task<OperationResult<IList<ApplicationComment>>> GetAllForApplicationAsync(User user, Guid applicationId, int page = 1, short pageSize = 100)
+        {
+            OperationResult<IList<ApplicationComment>> result = new ();
+            OperationResult<Application> applicationResult = await _applicationService.GetAsync(user, applicationId);
+            if (applicationResult.StatusCode == HttpStatusCode.OK && applicationResult.Result != null)
+            {
+                result.Result = await _commentRepository.GetAllForApplicationAsync(applicationId, page, pageSize);
+                result.StatusCode = HttpStatusCode.OK;
+            }
+            else
+            {
+                result.Messages.AddRange(applicationResult.Messages);
+            }
+
+            return result;
         }
     }
 }
